perf: narrow PokemonData numeric fields to column widths
PokemonData 592 -> 280 bytes (the design doc estimated 232; the
"pointer-carrying, last" field group's own 8-byte alignment requirement
reintroduces 7 bytes of padding a fully byte-count-monotonic order would
have avoided — see entity_sizes_test.go and pokemon.go's struct doc
comment for the measured breakdown). Pokemon 800 -> 456 bytes; still
carries changedFields and internal, both left for a later task.

Nullable numerics move from guregu/null's 16-byte wrappers to nulltypes
equivalents sized to the actual columns, and fields are reordered by
descending alignment.

Setters keep their null.X signatures and clamp out-of-range values to the
column boundary, counted by golbat_field_clamped_total. Direct field
assignments in pokemon_decode.go and weather_iv.go move to setter calls
(or, for AtkIv/DefIv/StaIv, direct clamp+assign inside calculateIv, which
remains their sole mutator by design) so they cannot skip the clamp.

Drops Capture1/2/3, which had setters but no callers and appeared in
neither pokemonSelectColumns nor pokemonBatchUpsertQuery. The webhook
payload keeps its capture_* fields at 0, which is what consumers have
always received.

The API response's float fields (weight/height/iv) widen through a
shortest-round-trip string rather than a naive float64 cast, so widening
a narrowed float32 back to *float64 doesn't expose float32 rounding
noise at float64 precision (float64(float32(3.14)) is 3.140000104904175,
not 3.14) - this keeps the golden JSON test's wire format byte-identical.

+++ decoder/pokemon_lure_test.go
@@ -101,7 +101,7 @@ func TestUpdateFromMapPlacesNewRecordFromCapturedFort(t *testing.T) {
 	if !pokemon.ExpireTimestampVerified {
 		t.Errorf("ExpireTimestampVerified = false, want true (GMO supplied ExpirationTimeMs)")
 	}
-	if got := pokemon.ExpireTimestamp.ValueOrZero(); got != expireMs/1000 {
+	if got := int64(pokemon.ExpireTimestamp.ValueOrZero()); got != expireMs/1000 {
 		t.Errorf("ExpireTimestamp = %d, want %d", got, expireMs/1000)
 	}
 	unlock()
@@ -136,9 +136,9 @@ func TestUpdateFromMapMergeAddsVerifiedExpiryOnce(t *testing.T) {
 	if !pokemon.updateFromMap(context.Background(), db.DbDetails{}, withExpiry, nil, "tester") {
 		t.Errorf("merge updateFromMap = false, want true (expiry contributed)")
 	}
-	if !pokemon.ExpireTimestampVerified || pokemon.ExpireTimestamp.ValueOrZero() != expireMs/1000 {
+	if !pokemon.ExpireTimestampVerified || int64(pokemon.ExpireTimestamp.ValueOrZero()) != expireMs/1000 {
 		t.Errorf("expiry = %d verified=%v, want %d verified=true",
-			pokemon.ExpireTimestamp.ValueOrZero(), pokemon.ExpireTimestampVerified, expireMs/1000)
+			int64(pokemon.ExpireTimestamp.ValueOrZero()), pokemon.ExpireTimestampVerified, expireMs/1000)
 	}
 
 	// Identical replay: nothing left to contribute.
@@ -258,7 +258,7 @@ func TestDiskEncounterFirstCreatesPlacedRecord(t *testing.T) {
 	if pokemon.ExpireTimestampVerified {
 		t.Errorf("ExpireTimestampVerified = true, want false (estimate)")
 	}
-	exp := pokemon.ExpireTimestamp.ValueOrZero()
+	exp := int64(pokemon.ExpireTimestamp.ValueOrZero())
 	if exp < before+lureSpawnLifetimeSeconds || exp > after+lureSpawnLifetimeSeconds {
 		t.Errorf("ExpireTimestamp = %d, want now+%ds (in [%d, %d])",
 			exp, lureSpawnLifetimeSeconds, before+lureSpawnLifetimeSeconds, after+lureSpawnLifetimeSeconds)
@@ -301,9 +301,9 @@ func TestGmoAfterDiskEncounterContributesVerifiedExpiry(t *testing.T) {
 	if pokemon == nil {
 		t.Fatalf("pokemon %d missing", encId)
 	}
-	if !pokemon.ExpireTimestampVerified || pokemon.ExpireTimestamp.ValueOrZero() != expireMs/1000 {
+	if !pokemon.ExpireTimestampVerified || int64(pokemon.ExpireTimestamp.ValueOrZero()) != expireMs/1000 {
 		t.Errorf("expiry = %d verified=%v, want %d verified=true",
-			pokemon.ExpireTimestamp.ValueOrZero(), pokemon.ExpireTimestampVerified, expireMs/1000)
+			int64(pokemon.ExpireTimestamp.ValueOrZero()), pokemon.ExpireTimestampVerified, expireMs/1000)
 	}
 	if got := pokemon.SeenType.ValueOrZero(); got != SeenType_LureEncounter {
 		t.Errorf("SeenType = %q, want %q (must not downgrade)", got, SeenType_LureEncounter)
@@ -338,9 +338,9 @@ func TestDiskEncounterAfterGmoUpgradesRecord(t *testing.T) {
 		t.Errorf("AtkIv = %d, want 15", got)
 	}
 	// GMO-verified expiry must survive: the estimate is only for new records.
-	if !pokemon.ExpireTimestampVerified || pokemon.ExpireTimestamp.ValueOrZero() != expireMs/1000 {
+	if !pokemon.ExpireTimestampVerified || int64(pokemon.ExpireTimestamp.ValueOrZero()) != expireMs/1000 {
 		t.Errorf("expiry = %d verified=%v, want %d verified=true (estimate must not overwrite)",
-			pokemon.ExpireTimestamp.ValueOrZero(), pokemon.ExpireTimestampVerified, expireMs/1000)
+			int64(pokemon.ExpireTimestamp.ValueOrZero()), pokemon.ExpireTimestampVerified, expireMs/1000)
 	}
 	unlock()
 }
